+++ docs/source/conf.py
@@ -18,11 +18,11 @@
 # -- Project information -----------------------------------------------------
 
 project = 'urlwatch'
-copyright = '2020, Thomas Perl'
+copyright = '2021, Thomas Perl'
 author = 'Thomas Perl'
 
 # The full version, including alpha/beta/rc tags
-release = '2.22'
+release = '2.23'
 
 
 # -- General configuration ---------------------------------------------------