+++ lib/urlwatch/html2txt.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of urlwatch (https://thp.io/2008/urlwatch/).
-# Copyright (c) 2008-2020 Thomas Perl <m@thp.io>
+# Copyright (c) 2008-2021 Thomas Perl <m@thp.io>
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -98,7 +98,7 @@ def html2text(data, baseurl, method, options):
 
     for k, v in options.items():
         if v is not None:
-            cmd.append('-{} {}'.format(k, v))
+            cmd.extend(['-{}'.format(k), '{}'.format(v)])
         else:
             cmd.append('-{}'.format(k))