+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 
 ## [UNRELEASED]
 
+## [2.23] -- 2021-04-10
+
+### Added
+
+- New filter: `pretty-xml` to indent/pretty-print XML documents
+- New filter: `jq` to parse, transform, and extract JSON data
+- New reporter: `prowl` (by nitz)
+
+### Fixed
+
+- Proper multi-line highlighting for wdiff (PR#615, by kongomongo)
+- Fix command-line generation for html2text (PR#619, by Eloy Paris)
+
 ## [2.22] -- 2020-12-19
 
 ### Added